% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annealing.R
\name{annealing_solver}
\alias{annealing_solver}
\title{ctor for annealing solver}
\usage{
annealing_solver(normalization = TRUE, schedule = c("fast", "boltzmann"),
  initial_temperature = 1, final_temperature = 1e-06, verbose = FALSE)
}
\arguments{
\item{initial_temperature}{initial temperature}

\item{final_temperature}{final temperature}

\item{normalize}{adjust all weights so that all possible changes have
mean square 1.0}

\item{boltzmann}{annealing or fast annealing}
}
\description{
ctor for annealing solver
}
