% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bionet_example}
\alias{bionet_example}
\title{Example MWCS instance obtained from BioNet package tutorial}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
bionet_example
}
\description{
Example MWCS instance obtained from BioNet package tutorial
}
\keyword{datasets}
