% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{budget<-}
\alias{budget<-}
\title{assignment operator for the budget}
\usage{
budget(x) <- value
}
\arguments{
\item{x}{a variable name.}

\item{value}{a value to be assigned to x.}
}
\description{
assignment operator for the budget
}
