% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{budget_costs<-}
\alias{budget_costs<-}
\title{assignment operator for budget costs}
\usage{
budget_costs(x) <- value
}
\arguments{
\item{x}{a variable name.}

\item{value}{a value to be assigned to x.}
}
\description{
assignment operator for budget costs
}
