% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{budgets<-}
\alias{budgets<-}
\title{assignment operator for budgets}
\usage{
budgets(x) <- value
}
\arguments{
\item{x}{a variable name.}

\item{value}{a value to be assigned to x.}
}
\description{
assignment operator for budgets
}
