% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{edge_weights<-}
\alias{edge_weights<-}
\title{assignment operator for edge weights}
\usage{
edge_weights(x) <- value
}
\arguments{
\item{x}{a variable name.}

\item{value}{a value to be assigned to x.}
}
\description{
assignment operator for edge weights
}
