% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{==.mwcs_instance}
\alias{==.mwcs_instance}
\title{equality operator for mwcs instances}
\usage{
\method{==}{mwcs_instance}(x, y)
}
\arguments{
\item{x, y}{mwcs instances}
}
\description{
equality operator for mwcs instances
}
