% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GAM}
\alias{GAM}
\title{GAM instances for MWCS problem}
\format{A vector of named vertex-weighted igraph instances}
\source{
\url{http://dimacs11.zib.de/instances/MWCS-GAM.zip}
}
\usage{
GAM
}
\description{
A dataset containing some real-world instances appeared in network
enrichment analysis tool
\href{https://artyomovlab.wustl.edu/shiny/gam/}{Shiny GAM}
}
\keyword{datasets}
