% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gam_example}
\alias{gam_example}
\title{GAM instance for MWCS problem}
\format{
A vector of named vertex-weighted igraph instances
}
\source{
\url{http://dimacs11.zib.de/instances/MWCS-GAM.zip}
}
\usage{
gam_example
}
\description{
A dataset containing some real-world instances appeared in network
enrichment analysis tool
\href{https://artyomovlab.wustl.edu/shiny/gam/}{Shiny GAM}
}
\keyword{datasets}
