% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gatom_example}
\alias{gatom_example}
\title{Example of graph from which an SGMWCS instance can be obtained}
\format{
An object of class \code{igraph} of length 194.
}
\usage{
gatom_example
}
\description{
The graph is based on \code{gatom} package
}
\keyword{datasets}
