% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{get_weight}
\alias{get_weight}
\title{Calculate weight of the solution. MWCS, GMWCS and SGMWCS instances are supported}
\usage{
get_weight(solution)
}
\arguments{
\item{solution}{Either \code{mwcsp_solution} or `igraph`` object representing the solution}
}
\value{
Weight of the subgraph
}
\description{
Calculate weight of the solution. MWCS, GMWCS and SGMWCS instances are supported
}
