% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmwcs_example}
\alias{gmwcs_example}
\title{Example GMWCS instance}
\format{
An object of class \code{igraph} of length 194.
}
\usage{
gmwcs_example
}
\description{
Instance is based on \code{gatom} package.
}
\keyword{datasets}
