% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmwcs_small_instance}
\alias{gmwcs_small_instance}
\title{Small example of GMWCS instance for demonstration purposes.}
\format{
An object of class \code{igraph} of length 5.
}
\usage{
gmwcs_small_instance
}
\description{
Small example of GMWCS instance for demonstration purposes.
}
\keyword{datasets}
