% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwcs.R
\name{gmwcs_solver}
\alias{gmwcs_solver}
\title{Returns a solver object for GMWCS solver}
\usage{
gmwcs_solver(cplex_bin, cplex_jar, gmwcs_jar,
  threads = parallel::detectCores(), timelimit = NULL, verbose = TRUE)
}
\arguments{
\item{cplex_bin}{a path to cplex library file}

\item{cplex_jar}{a path to cplex JNI jar file}

\item{gmwcs_jar}{a path to GMWCS solver jar file}

\item{threads}{number of threads for simultaneous computation}

\item{timelimit}{maximum number of seconds to solve the problem}

\item{verbose}{whether or not be verbose}
}
\description{
Returns a solver object for GMWCS solver
}
