% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{java_solver}
\alias{java_solver}
\title{ctor for java solver}
\usage{
java_solver(cplex_jar, cplex_bin, solver_jar, timelimit, root, threads,
  unrooted_time, rooted_time)
}
\description{
ctor for java solver
}
