% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{max_cardinality<-}
\alias{max_cardinality<-}
\title{assignment operator for the maximum cardinality of solution}
\usage{
max_cardinality(x) <- value
}
\arguments{
\item{x}{a variable name.}

\item{value}{a value to be assigned to x.}
}
\description{
assignment operator for the maximum cardinality of solution
}
