% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwcs_example}
\alias{mwcs_example}
\title{Example MWCS instance}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
mwcs_example
}
\description{
Instance is based on \code{gatom} package.
}
\keyword{datasets}
