% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwcs_small_instance}
\alias{mwcs_small_instance}
\title{Small example of MWCS instance for demonstration purposes.}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
mwcs_small_instance
}
\description{
Small example of MWCS instance for demonstration purposes.
}
\keyword{datasets}
