% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_solvers.R
\name{normalize_sgmwcs_instance}
\alias{normalize_sgmwcs_instance}
\title{Helper function to convert a graph into a proper SGMWCS instance}
\usage{
normalize_sgmwcs_instance(
  g,
  nodes.weight.column = "weight",
  edges.weight.column = "weight",
  nodes.group.by = "signal",
  edges.group.by = "signal",
  group.only.positive = TRUE
)
}
\description{
Helper function to convert a graph into a proper SGMWCS instance
}
