% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{parameters}
\alias{parameters}
\title{The method returns all parameters supported by specific solver}
\usage{
parameters(solver)
}
\arguments{
\item{solver}{a solver object}
}
\value{
A table containing parameter names and possible values for each parameter.
}
\description{
The method returns all parameters supported by specific solver
}
