% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{rmwcs}
\alias{rmwcs}
\title{Generates a rmwcs solver with corresponding parameters}
\usage{
rmwcs(timelimit = 1800L, max_iterations = 1000L, beta_iterations = 5L,
  separation = "strong", max_age = 10L, startcons = TRUE,
  pegging = TRUE, sep_iterations = 10L, sep_iter_freeze = 50L,
  heur_iterations = 10L, subgradient = "classic", beta = 2,
  verbose = FALSE)
}
\arguments{
\item{timelimit}{Timelimit in seconds}

\item{max_iterations}{Maximum number of iterations for Lagrangian solver}

\item{verbose}{Whether to print solving progress}
}
\description{
Generates a rmwcs solver with corresponding parameters
}
