% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{root<-}
\alias{root<-}
\title{assignment operator for the root}
\usage{
root(x) <- value
}
\arguments{
\item{x}{a variable name.}

\item{value}{a value to be assigned to x.}
}
\description{
assignment operator for the root
}
