% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{set_parameters}
\alias{set_parameters}
\title{Sets values of specific parameters}
\usage{
set_parameters(solver, ...)
}
\arguments{
\item{solver}{a solver}

\item{...}{listed parameter names and values assigned to them}
}
\value{
The solver with parameters changed.
}
\description{
Sets values of specific parameters
}
