% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgmwcs_example}
\alias{sgmwcs_example}
\title{Example SGMWCS instance}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
sgmwcs_example
}
\description{
Instance is based on \code{gatom} package.
}
\keyword{datasets}
