% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgmwcs_small_instance}
\alias{sgmwcs_small_instance}
\title{Small example of SGMWCS instance for demonstration purposes.}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
sgmwcs_small_instance
}
\description{
Small example of SGMWCS instance for demonstration purposes.
}
\keyword{datasets}
