% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{solution}
\alias{solution}
\title{Returns the most weighted connected subgraph}
\usage{
solution(instance)
}
\arguments{
\item{instance}{an instance of MWCS problem}
}
\description{
Returns the most weighted connected subgraph
}
