% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwcs.R, R/solvers.R
\name{solve_mwcsp.sgmwcs_solver}
\alias{solve_mwcsp.sgmwcs_solver}
\alias{solve_mwcsp.gmwcs_solver}
\alias{solve_mwcsp}
\title{Solves a MWCS instance}
\usage{
\method{solve_mwcsp}{sgmwcs_solver}(solver, instance, signals, ...)

\method{solve_mwcsp}{gmwcs_solver}(solver, instance, ...)

solve_mwcsp(solver, instance, ...)
}
\arguments{
\item{solver}{a solver object.}

\item{instance}{an MWCS instance.}

\item{signals}{kkk}

\item{...}{other arguments passed to other methods.}
}
\description{
Solves a MWCS instance
}
