% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmwcs.R
\name{solve_mwcsp.rmwcs_solver}
\alias{solve_mwcsp.rmwcs_solver}
\title{Solves MWCS problem using relax-and-cut approach}
\usage{
\method{solve_mwcsp}{rmwcs_solver}(solver, instance, max_cardinality = NULL, budget = NULL, ...)
}
\arguments{
\item{solver}{a solver object.}

\item{instance}{an MWCS instance.}

\item{max_cardinality}{integer maximum number of vertices in solution.}

\item{budget}{numeric maximum budget of solution.}

\item{...}{other arguments passed to other methods.}
}
\description{
Solves MWCS problem using relax-and-cut approach
}
