% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annealing.R
\name{solve_mwcsp.simulated_annealing_solver}
\alias{solve_mwcsp.simulated_annealing_solver}
\title{Solve generalized maximum weight subgraph problem using simulated annealing}
\usage{
\method{solve_mwcsp}{simulated_annealing_solver}(solver, instance, ...)
}
\arguments{
\item{solver}{An annealing solver object.}

\item{instance}{An igraph instance to work with.}

\item{...}{other arguments passed to other methods.}
}
\value{
An object of class mwcsp_solution.
}
\description{
Solve generalized maximum weight subgraph problem using simulated annealing
}
