% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{timelimit<-}
\alias{timelimit<-}
\title{Sets time limitation for a solver}
\usage{
timelimit(x) <- value
}
\arguments{
\item{x}{a variable name.}

\item{value}{a value to be assigned to x.}
}
\description{
Sets time limitation for a solver
}
