% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{upper_bound}
\alias{upper_bound}
\title{Returns the best known estimation on the upper bound of the score}
\usage{
upper_bound(instance)
}
\arguments{
\item{instance}{an instance of the MWCS problem}
}
\description{
Returns the best known estimation on the upper bound of the score
}
