% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annealing.R
\name{annealing_solver}
\alias{annealing_solver}
\title{ctor for annealing solver}
\usage{
annealing_solver(normalization = TRUE, schedule = c("fast",
  "boltzmann"), initial_temperature = 1, final_temperature = 1e-06,
  verbose = FALSE)
}
\arguments{
\item{normalization}{adjust all weights so that all possible changes have
mean square 1.0}

\item{schedule}{boltzmann annealing or fast annealing}

\item{initial_temperature}{initial value for the temperature}

\item{final_temperature}{final value for the temperature}

\item{verbose}{whether be verbose or not}
}
\description{
Simulated annealing is a heuristic method of solving optimization problems.
Typically, it allows to find some good solution in a short time. This
implementation doesn't compute any upper bound on solution, so there is no
guarantee of optimality of solution provided.
}
