% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{get_instance_type}
\alias{get_instance_type}
\title{Check the type and the validity of an MWCS instance}
\usage{
get_instance_type(instance)
}
\arguments{
\item{instance}{\code{igraph} object, containing an instance to be checked}
}
\value{
A list with members \code{type} containing the type of the instance,
\code{valid} -- boolean flag indicating whether the instance is valid or not,
\code{errors} -- a character vector containing the error messages

A list with two fields: the type of the instance with which it will
be treated by solve_mwcsp function and boolean showing validness of the instance.
}
\description{
Check the type and the validity of an MWCS instance
}
\examples{
data(mwcs_example)
get_instance_type(mwcs_example)
}
