% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwcs.R
\name{gmwcs_solver}
\alias{gmwcs_solver}
\alias{sgmwcs_solver}
\title{Construct a GMWCS solver}
\usage{
gmwcs_solver(cplex_bin, cplex_jar, threads = parallel::detectCores(),
  timelimit = NULL, memory = "2G", verbose = TRUE)

sgmwcs_solver(cplex_bin, cplex_jar, threads = parallel::detectCores(),
  timelimit = NULL, memory = "2G", verbose = TRUE)
}
\arguments{
\item{cplex_bin}{a path to cplex library file}

\item{cplex_jar}{a path to cplex JNI jar file}

\item{threads}{number of threads for simultaneous computation}

\item{timelimit}{maximum number of seconds to solve the problem}

\item{memory}{maximum amount of memory(-Xmx flag)}

\item{verbose}{whether or not be verbose}
}
\description{
This solver uses reformulation of MWCS problem in terms of mixed integer
programming. The later problem can be efficiently solved with
commercial optimization software. This solver uses CPLEX and requires
it to be installed.
}
