% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{mwcs_instance}
\alias{mwcs_instance}
\title{ctor for mwcs_instance}
\usage{
mwcs_instance(graph, parse_vertex_weights = TRUE,
  parse_edge_weights = FALSE, parse_budgets = FALSE)
}
\arguments{
\item{graph}{a graph}

\item{parse_vertex_weights}{wether or not parse vertex attribute "weight"}

\item{parse_edge_weights}{wether or not parse edge attribute "weight"}

\item{parse_budgets}{wether or not parse vertex attribute "budget"}
}
\description{
ctor for mwcs_instance
}
