% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{rmwcs}
\alias{rmwcs}
\title{Generates a rmwcs solver with corresponding parameters}
\usage{
rmwcs(timelimit = 1800L, max_iterations = 1000L, beta_iterations = 50L,
  separation = "strong", max_age = 1000L, startcons = TRUE,
  pegging = TRUE, sep_iterations = 1L, sep_iter_freeze = 1L,
  heur_iterations = 1L, subgradient = "classic", beta = 2,
  verbose = FALSE)
}
\arguments{
\item{timelimit}{Timelimit in seconds}

\item{max_iterations}{Maximum number of subgradient iterations}

\item{beta_iterations}{Number of nonimproving iterations until beta is halfed}

\item{separation}{Separation: "strong" of "fast"}

\item{max_age}{Extending the life of non-violated inequalities}

\item{startcons}{Whether to add flow-conservation/degree cons at start}

\item{pegging}{Pegging (variable fixing)}

\item{sep_iterations}{After how many iterations we are separating}

\item{sep_iter_freeze}{After how many iterations we are checking added ineqs}

\item{heur_iterations}{After how many iterations we are doing heuristics}

\item{subgradient}{Subgradient: "classic", "average", "cft"}

\item{beta}{Beta for subgradient}

\item{verbose}{Whether to print solving progress}
}
\description{
Generates a rmwcs solver with corresponding parameters
}
