% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{solve_mwcsp}
\alias{solve_mwcsp}
\title{Solves a MWCS instance}
\usage{
solve_mwcsp(solver, instance, ...)
}
\arguments{
\item{solver}{a solver object.}

\item{instance}{an MWCS instance.}

\item{...}{other arguments passed to other methods.}
}
\description{
Solves a MWCS instance
}
