% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scipjack.R, R/solvers.R, R/virgo.R
\name{solve_mwcsp.scipjack_solver}
\alias{solve_mwcsp.scipjack_solver}
\alias{solve_mwcsp}
\alias{solve_mwcsp.virgo_solver}
\title{Solves a MWCS instance}
\usage{
\method{solve_mwcsp}{scipjack_solver}(solver, instance, ...)

solve_mwcsp(solver, instance, ...)

\method{solve_mwcsp}{virgo_solver}(solver, instance, ...)
}
\arguments{
\item{solver}{a solver object.}

\item{instance}{an MWCS instance.}

\item{...}{other arguments passed to other methods.}
}
\description{
Solves a MWCS instance
}
